#582 build fix

+++ ui/viewmodel/wallet/tx_object_list.cpp
@@ -267,7 +267,7 @@ QVariant TxObjectList::data(const QModelIndex &index, int role) const
         case Roles::TimeCreatedSort:
             return static_cast<qulonglong>(value->timeCreated());
         case Roles::AmountGeneralSort:
-            return beamui::UIStringToAmount(value->getAmountGeneral());
+            return static_cast<qulonglong>(beamui::UIStringToAmount(value->getAmountGeneral()));
         case Roles::AmountGeneral:
             return value->getAmountGeneral();
         case Roles::Rate:
@@ -432,7 +432,7 @@ QVariant TxObjectList::data(const QModelIndex &index, int role) const
             return r;
         }
         case Roles::AmountSecondCurrencySort:
-            return beamui::UIStringToAmount(value->getAmountSecondCurrency());
+            return static_cast<qulonglong>(beamui::UIStringToAmount(value->getAmountSecondCurrency()));
         case Roles::AmountSecondCurrency:
             return value->getAmountSecondCurrency();
         case Roles::IsMultiAsset:
